cerbero: Only remove make_dir if it's not empty
It's not needed, and it fixes a race on Windows between deleting an
empty directory and checking if it exists with os.makedirs

+++ cerbero/build/build.py
@@ -468,7 +468,7 @@ class Meson (MakefilesBase):
         if not self.config_sh:
             raise Exception("The 'meson' build system was not found")
 
-        if os.path.exists(self.make_dir):
+        if os.path.exists(self.make_dir) and os.listdir(self.make_dir):
             shutil.rmtree(self.make_dir)
         os.makedirs(self.make_dir)
